requirements.txt
streamlit
pandas
openpyxl
sentence-transformers
torch
numpy
faiss-cpu
streamlit-searchbox